Track exceptions in Application Insights publisher
Previous version ignored reporting any exception contained in
reports. Most of the Xabaril health checks do send a value for
HealthCheckResult.Exception when the check fails, but they weren't
being reported to App Insights.

Note that there is lot of room for improvements on this class, so far
I've only added a simple additional method to track exceptions.