Use MySqlDataSource in `AddMySql` by default (#2096)
* Send a MySQL ping packet by default. Fixes #2031

Users can opt in to setting a command that will be executed on the server, but the default is now a more efficient ping.

* Add AddMySql overload for MySqlDataSource.

This pushes users towards good defaults: firstly registering a MySqlDataSource as a singleton in DI, then just calling .AddHealthChecks().AddMySql() with no other arguments to retrieve and use that same data source for health checks.

* Add package README for HealthChecks.MySql.

* Use new MySqlHealthCheckOptions API.

There are two constructors: one to initialise with a MySqlDataSource, and one with a connection string. This helps enforce correct uasge of the API by ensuring that exactly one is set.